+++ tests/infrastructures/doctrine/Recipe/Step/ODM/GetStreamFromMediaTest.php
@@ -21,6 +21,8 @@
  * @author      Richard Déloge <richarddeloge@gmail.com>
  */
 
+declare(strict_types=1);
+
 namespace Teknoo\Tests\East\Website\Doctrine\Recipe\Step\ODM;
 
 use Doctrine\ODM\MongoDB\Repository\GridFSRepository;